Remove all Bytes not directly facing a socket.
Not tested, on laptop :(

+++ backend/src/main/scala/com/cyborg/databaseIO.scala
@@ -8,8 +8,8 @@ import scala.language.higherKinds
 
 object DatabaseIO {
 
-  def arrayBreaker[F[_]](chunkSize: Int): Pipe[F,Array[Byte], Byte] = {
-    def get: Handle[F,Array[Byte]] => Pull[F,Byte,Unit] = h => {
+  def arrayBreaker[F[_]](chunkSize: Int): Pipe[F,Array[Int], Int] = {
+    def get: Handle[F,Array[Int]] => Pull[F,Int,Unit] = h => {
       h.receive1 {
         (v, h) => {
           println(" !! Breaking Array !! ")
@@ -18,7 +18,7 @@ object DatabaseIO {
         }
       }
     }
-    def unload(dudes: List[Array[Byte]]): Handle[F,Array[Byte]] => Pull[F,Byte,Unit] = h => {
+    def unload(dudes: List[Array[Int]]): Handle[F,Array[Int]] => Pull[F,Int,Unit] = h => {
       dudes match {
         case head :: t => Pull.output(Chunk.seq(head)) >> unload(t)(h)
         case _ => get(h)
@@ -28,19 +28,19 @@ object DatabaseIO {
   }
 
   // val filteredChannelStreams: Stream[Task,List[Stream[Task,Array[Byte]]]] = Stream.eval(filteredChannelStreamTask)
-  def databaseStream[F[_]: Async](fcs: Stream[F,List[Stream[F,Array[Int]]]]): Stream[F,Byte] = fcs flatMap (channelStreamList =>
+  def databaseStream[F[_]: Async](fcs: Stream[F,List[Stream[F,Array[Int]]]]): Stream[F,Int] = fcs flatMap (channelStreamList =>
     {
-      val unpacked: List[Stream[F, Vector[Byte]]] =
+      val unpacked: List[Stream[F, Vector[Int]]] =
         channelStreamList.map(_.through(arrayBreaker(512)).through(utilz.vectorize(1024*4)))
 
-      val spliced: Stream[F, Vector[Byte]] =
-        (Stream[F, Vector[Byte]](Vector[Byte]()).repeat /: unpacked){
-          (b: Stream[F, Vector[Byte]], a: Stream[F, Vector[Byte]]) => b.zipWith(a){
+      val spliced: Stream[F, Vector[Int]] =
+        (Stream[F, Vector[Int]](Vector[Int]()).repeat /: unpacked){
+          (b: Stream[F, Vector[Int]], a: Stream[F, Vector[Int]]) => b.zipWith(a){
             (λ, µ) => µ ++: λ
           }
         }
 
-      val rechunked: Stream[F, Byte] =
+      val rechunked: Stream[F, Int] =
         spliced.through(utilz.chunkify)
 
       rechunked
